+++ module/model/DOC.html
@@ -213,6 +213,23 @@ Some implementations also allow for the ability to read/write binary data as <a 
    * @param location The location of the blob
    */</span>
   <span class="token function">updateBlobMeta</span><span class="token punctuation">(</span>location<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">,</span> meta<span class="token operator">:</span> BlobMeta<span class="token punctuation">)</span><span class="token operator">:</span> <span class="token builtin">Promise</span><span class="token operator">&lt;</span><span class="token keyword">void</span><span class="token operator">></span><span class="token punctuation">;</span>
+
+  <span class="token comment">/**
+   * Produces an externally usable URL for sharing limited read access to a specific resource
+   *
+   * @param location The asset location to read from
+   * @param exp Expiry
+   */</span>
+  getBlobReadUrl<span class="token operator">?</span><span class="token punctuation">(</span>location<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">,</span> exp<span class="token operator">?</span><span class="token operator">:</span> TimeSpan<span class="token punctuation">)</span><span class="token operator">:</span> <span class="token builtin">Promise</span><span class="token operator">&lt;</span><span class="token builtin">string</span><span class="token operator">></span><span class="token punctuation">;</span>
+
+  <span class="token comment">/**
+   * Produces an externally usable URL for sharing allowing direct write access
+   *
+   * @param location The asset location to write to
+   * @param meta The metadata to associate with the final asset
+   * @param exp Expiry
+   */</span>
+  getBlobWriteUrl<span class="token operator">?</span><span class="token punctuation">(</span>location<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">,</span> meta<span class="token operator">:</span> BlobMeta<span class="token punctuation">,</span> exp<span class="token operator">?</span><span class="token operator">:</span> TimeSpan<span class="token punctuation">)</span><span class="token operator">:</span> <span class="token builtin">Promise</span><span class="token operator">&lt;</span><span class="token builtin">string</span><span class="token operator">></span><span class="token punctuation">;</span>
 <span class="token punctuation">{{'}'}}</span></code></pre>
   </figure>
