+++ module/model/src/service/blob.ts
@@ -1,4 +1,4 @@
-import { BinaryInput, BlobMeta, ByteRange } from '@travetto/runtime';
+import { BinaryInput, BlobMeta, ByteRange, TimeSpan } from '@travetto/runtime';
 
 /**
  * Support for Blobs CRUD.
@@ -39,4 +39,21 @@ export interface ModelBlobSupport {
    * @param location The location of the blob
    */
   updateBlobMeta(location: string, meta: BlobMeta): Promise<void>;
+
+  /**
+   * Produces an externally usable URL for sharing limited read access to a specific resource
+   *
+   * @param location The asset location to read from
+   * @param exp Expiry
+   */
+  getBlobReadUrl?(location: string, exp?: TimeSpan): Promise<string>;
+
+  /**
+   * Produces an externally usable URL for sharing allowing direct write access
+   *
+   * @param location The asset location to write to
+   * @param meta The metadata to associate with the final asset
+   * @param exp Expiry
+   */
+  getBlobWriteUrl?(location: string, meta: BlobMeta, exp?: TimeSpan): Promise<string>;
 }
