+++ related/travetto.github.io/src/app/documentation/gen/model/model.component.html
@@ -213,6 +213,23 @@ Some implementations also allow for the ability to read/write binary data as <a 
    * &#64;param location The location of the blob
    */</span>
   <span class="token function">updateBlobMeta</span><span class="token punctuation">(</span>location<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">,</span> meta<span class="token operator">:</span> BlobMeta<span class="token punctuation">)</span><span class="token operator">:</span> <span class="token builtin">Promise</span><span class="token operator">&lt;</span><span class="token keyword">void</span><span class="token operator">></span><span class="token punctuation">;</span>
+
+  <span class="token comment">/**
+   * Produces an externally usable URL for sharing limited read access to a specific resource
+   *
+   * &#64;param location The asset location to read from
+   * &#64;param exp Expiry
+   */</span>
+  getBlobReadUrl<span class="token operator">?</span><span class="token punctuation">(</span>location<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">,</span> exp<span class="token operator">?</span><span class="token operator">:</span> TimeSpan<span class="token punctuation">)</span><span class="token operator">:</span> <span class="token builtin">Promise</span><span class="token operator">&lt;</span><span class="token builtin">string</span><span class="token operator">></span><span class="token punctuation">;</span>
+
+  <span class="token comment">/**
+   * Produces an externally usable URL for sharing allowing direct write access
+   *
+   * &#64;param location The asset location to write to
+   * &#64;param meta The metadata to associate with the final asset
+   * &#64;param exp Expiry
+   */</span>
+  getBlobWriteUrl<span class="token operator">?</span><span class="token punctuation">(</span>location<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">,</span> meta<span class="token operator">:</span> BlobMeta<span class="token punctuation">,</span> exp<span class="token operator">?</span><span class="token operator">:</span> TimeSpan<span class="token punctuation">)</span><span class="token operator">:</span> <span class="token builtin">Promise</span><span class="token operator">&lt;</span><span class="token builtin">string</span><span class="token operator">></span><span class="token punctuation">;</span>
 <span class="token punctuation">{{'}'}}</span></code></pre>
   </figure>
